Add ArticleClient component for improved article data fetching; refactor Article page to utilize new component and streamline locale handling.
